Refactor files in command package for test cases
